+++ tests/lint/hardcoded-absolute-path.pas
@@ -1,0 +1,15 @@
+unit Paths;
+
+interface
+
+implementation
+
+procedure P;
+var
+  S: string;
+begin
+  S := 'C:\Temp\log.txt';
+  S := 'relative\path';
+end;
+
+end.