+++ tests/lint/win64-pointer-cast.pas
@@ -1,0 +1,15 @@
+unit W64;
+
+interface
+
+implementation
+
+procedure P(Ptr: Pointer; I: Integer);
+var
+  N: Integer;
+begin
+  N := Integer(Ptr);
+  N := Integer(I);
+end;
+
+end.